Bug 1712578 - Use TMPDIR if defined over /tmp in shell scripts. a=bustage
We have some logic that directly references /tmp that assumes it will have
bountiful storage available, but the reality is that it does not.  In
particular, `crossref.sh` was directly sorting the ~1.6G into the /tmp dir
which is very close to the amount of free space on our `/` EBS image.

This is not meant to be the end state fix for this, so I've just corrected
the shell scripts where it's trivial to evaluate an environment variable
and failover.  I have not touched the following other files under /scripts
that have the same problem but involve python/JS:
- output-dir.js
- idl-analyze.py
- idl-analyze.sh (because of the above)
- build-codesearch.py

The simplest solution in terms of avoiding using the wrong tmp dir would
be to just make sure that the /tmp dir is actually backed by the local
SSD, but that would ideally want to happen early in the machine boot
process to avoid /tmp already being in use when we actually get around to
creating the scratch drive.

Maybe that's okay though?  Otherwise we'll potentially want to enforce
some kind of invariant about TMPDIR being set and/or the indexing process
having a "tmp" directory in each index directory (which would make sense
for the purposes of having `mv` operating within the same filesystem).