Bug 1641372 - router.py python3 drift fixes
The fancy branch UI work all built on the "sorch" search endpoint which broke
with the change to python3 but I didn't notice it when rebasing because the
intent was not to surface any of the fancy UI and so it didn't get tested
(and there currently is not test coverage for router.py).

These changes are to account for changes to urlparse that were made by six
and the minor method change introducing `generateJson`.

With these fixes it should now be possible to run (outside of the VM) on a
machine with the VM running the following and see something almost useful:

```
curl -H "Accept: application/json" 'http://localhost:16995/tests/sorch?q=superhero&path=' | jq .
```