Bug 1729178 - Enable serde_json preserve_order feature
By default serde-json re-orders objects alphabetically because it uses a
BTreeMap under the hood.  By enabling "preserve_order" an IndexMap is
used instead which maintains the ordering.

This commit enables preserve_order because router.py emits an explicitly
ordered object dictionary and so we need that order reflected in testing.

This results in changes to all the test snapshots, so this commit only
makes this change and commits those changes.

This will have a corresponding change in mozsearch-mozilla very shortly.