Bug 1728000 - Update try build and testing docs.
While writing the testing docs I realized there was a gap for the release
versus non-release variations, so I put an XXX in the document and want
to both address that and then update the docs once it's addressed.  I
think I may have some stashed changes to dig out that may contain the
fix needed.

I'm landing this now because I'm unlikely to have time to directly work
on this during the work week and want this information to be available
in case someone else can dig into any enhancements.