Bug 1732791 - Update git-cinnabar to get rid of fsck warnings.
This isn't currently on the release branch but that's not a problem.

We are leaving existing fsck runs in place on purpose, we just don't
want the spurious warnings that happen from the 7-day warning interval
and the inherent phase issues of a 7-day warning when we do checks
every 7 days.

Cinnabar has some experimental support for Python 3.x but I'm not
enabling that.  I have tried to clarify the 2.7 dep in the indexer case
because I'm removing the (redundant) python2.7 install from the AWS
provision script with the removal of the cloudwatch support in the next
commit.