Bug 1737298 - document how to update checks on the indexer
This is the cheat sheet I made as I went so that one can login to the
indexer and review the changes and then sync them locally to create a
pull request.

Longer term I just want the indexer to continue running and stand up a
web-server on a special failure channel so that things can be done
locally, but it's handy to have this documented in the interim.