Dark mode.
Simpler, hopefully more palatable implementation of dark mode than #415.

The main trick that allows this to be so tiny is that we rely on the
browser's form controls (using color-scheme: dark).

The other main trick is including the copy SVG icons in the source
directly, so that we don't have to have multiple versions of them and
they can just use `fill: currentColor`.