Bug 1744153 - stuck line box-shadow wants padding
The box-shadow for stuck lines makes it hard to read the directly linked lines.

This patch adds an additional line of scroll padding so that the box-shadow can't possibly
interfere with the line.

The one caveat to this fix is that when directly linking to the first level of nesting, like
`namespace mozilla {`, we will add an extra line of padding.  This is because we are
choosing to favor the case where there's 2 stuck lines (like a nested namespace) and
avoid adding another selector or crazy math.  If only calc supported a
[signum function](https://en.wikipedia.org/wiki/Sign_function)!