Correct WithLocation JSON serialization misuse.
A nuance I'd missed was that there were cases where we have
`WithLocation<Vec<T>>` as an aggregate where the JSON serialization
then produces a separate JSON object for each item in `Vec<T>`.  This
doesn't work with a bidirectional JSON mapping and so I've changed the
merger logic to manually unroll this scenario.

The test case that helped validate this gets its own series of commits
in 2 steps:
1. Add the merge usage with the file present a single time.
2. Add the second usage.