Bug 1702916 - Switch to binary-searched crossref usage with external blobs
Note that in the change to use ustr as part of the conversion to serde_json
(in order to reduce manual work to perform reference-counted string
interning) there was a regression and `BTreeMap<Ustr, ...>` did not
actually provide a lexicographically-sorted iterator, but instead an
iterator over when strings were interned.  This hadn't mattered at the
time because `crossrefs.py` never actually depended on `crossref` being
sorted; it built its own dictionary mapping symbols to offsets/lengths.

Our change to performing a binary search obviously ran afoul of this
and so we're switching to a patched version of Ustr that has a
corrected implementation of `Ord` and revised tests.

A side effect of correcting this regression is that the output from the
merge-analyses test has changed because we emit structured symbols in
BTreeMap order and this order is now lexicographic rather than its
effectively nonsensical ordering.  We don't actually care about the
order of lines in analysis files at this time and merging has always
permuted the lines, so this isn't meaningful other than producing a
hard-to-read diff.