searchfox-tool: Better CLI experience for 0 args or too many args.
The logic was hard-coded to assume a single arg and would panic if no
args were provided.  We now coerce the 0-args case to be "--help" and
try and prefix that with the explanation about the internal parsing.
That explanation is unlikely to make sense to anyone but me as the
author of this approach, but the >1 case will now error out with only
error text which will probably make sense to people in that context,
although possibly too late to save them from doing the escaping wrong.