Bug 1755807 - Add XPIDL Test Coverage and Fix XPIDL analysis.
== Test Coverage

Add XPIDL data to the test corpus that's partially extracted from
mozilla-central and can be updated via the new
`refresh-mozilla-central-stuff` script from a running config1 web
server.

The rationale behind this approach is that the XPIDL generation process
is non-trivial and likely to break if directly embedded in this repo,
so we just copy over the byproducts that we care about.  These will
also form the basis for our production m-c checks, which should ideally
make it easier to keep up-to-date because if one of them breaks, they
both should break (at least after running
`refresh-mozilla-central-stuff`.

== The Problem and Fix

Our source records for IDL files did not include "syntax" keys/values,
and these are necessary for the serde-based serialization to decide
that a source record is valid.

Our target records for IDL files were also missing "pretty" fields
which didn't really matter but is necessary for the `filter-analysis`
searchfox-tool command to be able to do an `-i` identifier match on
things, so it now does that too.

This provides for semantically highlighting the symbol in source files
and providing a "search" option.  It also provides a very useless "go
to definition" option which is the C++ definition which no one would
ever want.  But I think that's largely a longstanding issue and I have
made various exciting proposals in bug 1727789 that I think likely is
how this will be addressed, although maybe we can hack up a specialized
fix for XPIDL for this case.