Bug 1759064 - Multiple source records for IDL stuff.
This implements the core feature enhancement of providing source
records for the C++-specific XPIDL paths.  Test additions for
attributes that verify the attribute behavior come next in the
stack.

Note that this does show that we are failing to de-duplicate symbols
when emitting the HTML but I'm not fixing that in this patch because
I believe it's probably harmless.  If it's not harmless, it gets to
get fixed in another patch!