Bug 1760832 - Propagate env vars to sudo invocations directly.
Previously the environment variables were only added to the little
script but not propagated into the sudo invocation because sudo was not
being told to propagate any env vars.  For the sake of clarity and to
avoid any weird precedence issues, we now explicitly list the env vars
on each sudo invocation.