Bug 1764957 - Update git-cinnabar to the release revision
`git-cinnabar download` was failing because of a (presumably) expired
taskcluster artifact.  We've been using specific revisions to get fixes
and enhancements before they're released, but everything we've wanted
is now already on the `release` branch so we can just check that out
directly.

git-cinnabar is a(n AMAZING) foundational tool for us, so it makes
sense to at least always be on the most recent released version unless
we need to be on something more recent than release.  We're not
actually trying to pin our provisioning at specific revisions for tools
like git-cinnabar.