Bug 1728376 - Late-follow up heuristic limit change
Nika ran into an odd gap in `nsIAsyncInputStream::Available` results which turned out to be due to this limit.

I believe my general concerns previously were:
- router.py should not OOM / take forever if someone searches for nsISupports
- Adding more than 50 symbols potentially grows the result set too much.

Unfortunately, when we decide not to merge, we don't currently emit a marker that says "hey, there were some other potential results here, but we didn't show them" which means that this limit is a real hard cliff in terms of user experience.  I'd been hoping to do more follow-ups shortly after this, but ended up running out of professional time to work on this, and my personal efforts have more been on foundational rust-based work.  So I'm doubling the limit for now and will plan to improve these use-cases with the rust-based query work I'm doing.