Bug 1764957 - idempotently check out mozsearch for email purposes
In 747998194e3526085c7def0e986479b9c8dd36a2 I added a clone at the top
of both provisioner scripts but didn't deal with the reality that we
run both scripts for VM provisioning and that `git clone` will error if
you try and clone into an existing dir, even if the result would be the
same.  So we do a dir-check first.  (Also note that I think the update
steps at the bottom of each script would have created mozsearch too, so
it's really the lack of a check that's the problem.)

xref https://github.com/mozsearch/mozsearch/pull/501 which landed that
commit.