Bug 1762817 - Query chewing now builds graph rep, but not the pipelines.
This patch expands the TOML configuration file to be able to specify
graph relationships amongst groups based on the groups that are created
by the query expansions.

The logic is also expanded so that we populate the
QueryPipelineGroupBuilder rep which is what we dump in the tests.
However the actual pipeline is not yet built/run.  A few more pipeline
commands are necessary to make this work for the classic searchfox
experience, but we already have everything needed for SVG graphing, so
I'm going to start with that use-case.