The indexer now needs livegrep gRPC built too.
Now that our indexer checks run with the livegrep server running as
started by codesearch.py, we need to replicate the steps to build the
python livegrep gRPC bindings.  This was missed in local testing
because the VM runs both provisioner scripts.

This is a fixup to https://github.com/mozsearch/mozsearch/pull/505