Bug 1567724 - save off the output.sh parallel metadata for analysis
Without this change, all of the data about what GNU parallel gets up to
goes in the TMPDIR which makes it hard to figure out why things take so
long after the fact.  This puts all that information in `diags/output`.

The resulting `.par` files also provide out println!'s from
output-file.rs and I'm going to add more to this in an additional commit.