Bug 1567724 - Only load the desired tree for perf reasons.
It can be quite expensive to load each tree in the config file,
especially when `need_indexes` is true, like in `output-file.rs`, to
the tune of an extra 82 seconds of loading time.  This change makes
us only load the desired tree for commands that only operate on a
single tree at a time.  The notable exception is `web-server.rs`
which inherently operates on all trees.