Bug 1771620 - Pass paths to output-file over stdin
We can eliminate a massive amount of process startup overhead by having
output-file receive the list of files to process over stdin.  With
these changes we basically run one output-file instance per core.  The
main trade-off is that because we distribute the list of files up front
and don't do any dynamic scheduling, there's a potential for unequal
distribution of work.

Should this end up being a meaningful problem, we can:
- Randomize the contents of `all-files` to avoid clustering of larger
  files or ramifications of different path-lengths in the tree (since
  divvying is based on bytes, not lines).
- Ditch parallel in favor of just having output-file internally be
  multi-threaded and use a work-queue mechanism that can be aware of
  how much work each thread still has to do / does work-stealing / etc.