Bug 1773767 - Update the IPDL parser and add limited test coverage.
Because there's a lot of moving parts for the mozilla-central IPC
implementation, we just copy analysis files from a web-server that
has the C++ analysis data we want for the source files we want to show.

This provides ipdl-analyze.rs with the analysis files it needs to
scrape to establish the C++ symbol mapping.  And this is then enough
for the test repository to give us a full fidelity experience of
viewing the .ipdl file and sufficient fidelity when clicking on those
associated Send/Recv methods as they relate to the IPDL file.  That's
the end of our fidelity, although if we pulled down other source files
and their associated analysis files, we could get a little more.

The goal is to allow rapid iteration on the IPDL mapping and analysis
record generation without having to go through a full mozilla-central
lookup.

Minimal test checks added.