Bug 1759071 - Add tc-coverage-timestamps.sh
This is a helper script that previously lived just as a gist, but which
should really be in-tree.  I've updated the comment block which
explains how to install it because it bit-rotted somewhat.  (I
discovered this because of my own Ubuntu 22.04 upgrade and the inherent
Python upgrade that went along with it.)