Bug 1779672 - log codesearch stats, prevent default submit
codesearch has stats we can log as well, noting that the repr will elide
zero values.

Potentially more notably, we inhibit the default submit behavior of the
search field in favor of our magic AJAX logic.  It would potentially be
better to explicitly handle the user pressing enter as an acceleration
of `startSearchSoon` via `startSearch` since startSearchSoon has a
300ms delay, but I'm going to have that be an enhancement since this
minimal fix is easier and an overall win for the user and the server.