Bug 1703115 - Have web-server use exact git revisions indexer used.
Before this change, the web-server would use the same decision process
for which revisions of the git repositories to use as the indexer, but
since new patches can land during the hour(s) between the indexer
starting and the web-server starting, this has been a recipe for
regressions.  By having the indexer propagate the exact revisions it
used to the web-server, this addresses the problem.

This will require a re-provision because changes to
web-server-provision.sh to impact the web-server's update.sh script
which inherently requires re-provisioning.