Bug 1783761 - Upgrade to clap v3 from clap v2/structopt
We were slow enough in this upgrade that the migration ended up being
slightly more involved to make the deprecation warnings go away.  This
mainly entailed adding "value_parser" to everything.

This requires a re-provisioning because of our cargo deps policy that
we want the VM to already have cached the crates locally.