Bug 1786216 - fix current indexed revision permalink optimization
Python f-strings render byte strings of the form `b'foo'` exactly like
they are written in the source when rendered to string, and we were
unintentionally using `bytes` from our subprocess invocation.  This was
not previously a problem when we were using the `% fmt` mode of
formatting, but became a problem from f-strings' differing behavior.