Bug 1786216 - Further fix permalink current revision opt
The permalink optimization may actually have been broken even longer
than expected as by fixing the route, I revealed that we hadn't
compensated for the gzip optimization I'd done a long, long, long time
ago and so we were serving the 0-length files.  This implies the
format string might actually have had the same behavior for the byte
string and that this was broken since the change to python3.