Bug 1789329 - Add explicit mjs parsing, fallback module parsing, tests.
For JS parsing:
- If the file ends in ".mjs" we treat it as a module only.
- If the file ends in ".js" we try parsing it as a script and fall back
  to also trying it as a module if it doesn't parse as a script because
  imports/exports are illegal.

For the analysis:
- We process ExportDeclaration nodes by recursively processing the
  declaration it wraps.
- We otherwise ignore everything having to do with modules because our
  SCIP support should let us do clever things soon so it's a bit of a
  waste to try and do something smart since it will all just get thrown
  away.

I add some very basic definition checks to ensure that we've parsed the
modules and the export declarations.  I don't add any other
pre-existing JS checks at this time.  This is largely because we
haven't really seen regressions and also because we are 100% going to
have to change the symbol mapping when we move to SCIP.