Bug 1792322 - Treat JS scripts as modules on export decls too.
I special-cased us to re-parse when an import was encountered, but
should have also made this be the case for exports too!  I am going
to skip adding a check for this at this moment because of the
short-to-medium-term JS analysis changes coming and because this
additional logic is very straightforwardly related to the import
logic that had a check added.