Bug 1792322 - Add a bunch more downgrade heuristics.
Details in the comments for each heuristic, again noting that in
general, I do have some WIP that will let us ideally ingest
.eslintignore or the output of the eslint logic which is the
general fix for situations like this.

I'm not blocking this on that just because there's already so much
going on with that and there are potentially a bunch of logistical
issues related to the eslint processing and I don't want to block
that stack on such a potentially large undertaking.