Bug 1783761 - Clean up templates, footer should be part of scrolling area
A decision to try and leave each template in full control of the
scrolling div and the content area inside it needed us to introduce a
`scroll_footer.liquid` sub-template so that it could be part of the
scrolling area rather than have it be fixed (which was an oversight).

I've also cleaned up the footer link for filing a bug which was overdue
for removing the jokey part of the link text.  (Bugzilla's duplicate
detection is good and we don't have a problem with duplicate bugs being
filed; even when things overlap we get useful new examples.)