Bug 1656010 - Learn to ingest WPT manifest.json files.
Adds support for ingesting "wpt-mainfest.json" which is going to be
twinned for the "mozilla" separate WPT root.  This generalizes the
"nesting" type of "hierarchical-dict-dirs-are-dicts-files-are-values"
which we were already using for bugzilla-components ingestion but we
no longer require the lookup mechanism be used against the values.

Thanks to care previously taken, overlapping uses of
"concise.info.wpt.object" in the different WPT JSON files merge results
rather than clobbering them.