Bug 1763005 - Improve tracing logging to be cross-thread.
The initial explanation mechanism used the "set this subscriber to be
used on this thread only" using a guard for lifecycle purposes.  This
provide a linearized JSON log that wouldn't contaminate other threads,
but was inherently single-threaded because after the tokio 0.2 overhaul,
tokio didn't really have a good way to propagate that subscriber to
other threads along with the span.

Notionally
https://docs.rs/tracing-futures/0.2.5/tracing_futures/trait.WithSubscriber.html
would provide this, but as explained at
https://github.com/tokio-rs/tracing/issues/593 and also somewhat at
https://github.com/tokio-rs/tracing/discussions/1626 which gave rise
to our initial implementation (or a related issue did?), that doesn't
work currently.

In the bigger picture, we also do potentially want to do more than
per-"query" logging.  That said our use cases are strongly biased to
"here's this top-level span and I want a hierarchy of everything that
happened in this span and I'm going to report this back to the user or
the test case" as opposed to I want an omniscient view of everything
going on right now".  So something global was always desired.

The approach I've gone with is to discover tracing-forest and decide
that it's great for our next step, and maybe forever.  The primary
trade-off is that it buffers the span tree until it's ready to be
flushed but there are potentially a bunch of up-sides to this and also:
- Understanding what's going on in the system is most important.
- Our logging implemented thus far already has a concept of having the
  request explicitly be told that we want its logging, and this seems
  like a reasonable approach to continue.  Especially because I think
  it's likely we'll also want to support a "needle" mechanism where we
  conditionalize either the log or its output on being related to the
  needle.
- We're already throwing a fair bit of RAM at searchfox and both for
  performance reasons and just human ability to comprehend things, the
  "needle" mechanism I describe above likely means that our 80% case
  for the logging would be trying very hard to not exceed a megabyte
  of logs.
- tracing-forest explicitly has the concept of "a worker should
  process the logs" which is well suited for being able to use
  parallelism to our advantage in the processing and also to leverage
  the abstract span reps to potentially do log decimation on that
  worker thread.

All that said, I do expect to learn a bunch from this, but the most
important thing is that I can tell people how to run a query in explain
mode and have them see something reasonable and non-overwhelming and
am very willing to take a hit to efficiency in this opt-in logging
mode.

The main lesson learned in this whole experience is that things will
break badly if you try and hold an entered span across an await()
call.  The tracing docs are clear that you shouldn't do this, and
in some cases the type system will prevent you from doing this
(maybe the problem was I prefixed the entered guard with a `_`?),
but if we mess this up, it can absolutely cause tracing-forest to
be unable to close out the span hierarchy because a reference to a
span can end up stuck in a thread's thread-local-storage resulting
in `on_close` never being invoked for the span.