Bug 1801960 - Widen the context-menu word regexp to include #.
As :saschanaz pointed out, the context-menu's "search for the substring"
option is excluding the leading "#" for private symbols still, which is
weird.

This fix naively just widens the regex.  I imagine there could be regressions
for this and it could make sense for this JS to instead understand the
difference between:
- The user clicking on a symbol for which we can know the exact symbol string
  to use.
- The user clicking on random text which is not marked up as a symbol and for
  which we should use a regex heuristic.
- The user selecting a region of text and then clicking on it with the intent
  to potentially provide our context menu to search for that string, but where
  we will currently erase the user's own selection.