Bug 1656010 - Add probing mechanism, fix dict_eval to pass-in existing key value.
The new logging was invaluable here; the problem was that 2 separate
files were both trying to contribute to the "wpt" dictionary key, with
the second file clobbering the result of the first file.  Very much a
case where the thing I was debugging was not where the bug was!

The "probing" mechanism can be activated by running
`export PROBE_PATH=test_ima_disabled_test` prior to running
`make build-test-repo` and will result in a markdown file being written
to `diags/crossref/repo_ingestion.md` under the index root, which in
this case (in the VM) is `~/index/tests`.  The log will only have
evaluation information for files which match the regexp, which avoids
having to fish through the entire log.

Note that right now, the `repo_ingestion_explainer.liquid` template is
basically the same as the `pipeline_explainer.liquid` file, which is to
say that it JSONs what it is given.  Unfortunately, the span names come
after the nodes (usually?), so it is not fully optimal, but the plan is
also to not actually just dump JSON in the future.

I am very happy with how this has turned out, as while I'm not super
crazy about have invented a new processing mechanism for the data
ingestion here, the combination limited CISC-like primitives with the
granular tracing instrumentation (that gets aggressively constrained
by the probe mechanism) is arguably a lot better than would be
accomplished by requiring custom/ad-hoc jq or python scripts for
everything.