Bug 1656010 - Strip WPT path root off the null-inferred test ids.
The test IDs for the null WPT path case were using the full tree path
when they want to only have the WPT root path.  I've addressed this by
having the config explicitly strip off the path via
`strip_prefix_or_empty`, but arguably it could make sense for the
context to have a concept of `tree_path` and `rel_path` where the
`rel_path` would be relative to the `path_prefix`.  I've deferred this
for now and probably would want to wait to decide on this until we've
added some more files to ingest and seen what other capabilities we
might need.