Bug 1656010 - List WPT test ids in info box, link WPT and interop dashboards there.
Now that we have the test IDs for web platform tests, we can properly
link to the wpt.fyi "dashboard" and jgraham's interop dashboard.
Previously the WPT dashboard was in the navigation panel, but that
doesn't scale with the number of test IDs, and in particular, it
would not be obvious which file is which.

Currently I've laid out the links in the flow under a ul/li, but
visually it seems like we either want a table or we want to hide the
links in context menus.  An argument in favor of a table would be that
we could add a column that explains what global the test runs in before
the test ID, as at least as far as I am aware, most of the time that's
the big difference and arguably for a screen reader user that would be
preferable.