Bug 1703656 - Implement a settings page, minor context menu regression fix.
There was a little hiccup in bug 1809987 where it turns out the context
menu actually thought it was displaying icons, and with the CSS class
changes, the layout of the context menu regressed.  I hadn't realized
icons went there because back in 2021 we deleted the CSS glue that used
to always display search icons in the context menu.  (And I mean
always!  DXR used icons to differentiate between different classes of
options but in searchfox it seems we've only ever used the search
icon.  There is more work to do on this moving forward.)