Bug 1703656 - Remove settings page text detritus, bind feature qualities.
I had missed a final cleanup pass on the settings page text, leaving a
number of fragments of text that were obsoleted by the bullet
structuring.  This should now be addressed (hopefully)!  But there's
still conversations to have around the text which could lead to
future iteration.

I also realized the quality levels of features was not being reflected
into the page, so I added a mechanism for that by using identifiers
of the form "quality--group-name--key-name" where the
"group-name--key-name" following "quality--" is the same identifier we
expect to see when performing form binding.