Bug 1459188 - Don't trim whitespace when showing context, more efficient DOM
When using the secret `context:N` feature, we were truncating the whitespace
off the front of the actual hit line, while retaining the leading whitespace
for the context lines, which resulted in a jarring experience.  This change
avoids the normalization if we have context available.

While looking at this, I noticed that we were also wastefully prefixing all
the context lines with `<b></b>` because they had no before or middle, but
they did have an after, and before the introduciton of the context case,
this was not a situation that would happen.