Bug 1812838 - Add a trees subdirectory that gets exposed into docker.
In order to index local trees, it's desirable to have a means of adding
symlinks which will not be added to the repo (hence the .gitignore
directives), but can be indexed.  We teach run-docker.sh how to do some
accidentally working automatic mounting.  The container will need to be
manually pruned when new symlinks are added, I think.  Or maybe we
could add an idempotent pass if the container is stopped.  Future work.