Bug 1812838 - More "trees" support, scip-typescript/python provisioning.
The trees target helps experimentation with indexing m-c subtrees with
scip-typescript without interfering with our continued use of js-analyze
and scip-python as well.  (scip-python doesn't interfere with anything,
but it's not useful yet either.)

To this end, the provisioning script is updated to install rtx and use
it to get a modern node.js and use it to install scip-typescript and
scip-python.