Provision scip CLI 0.3.0 to get `scip print`
`scip print` is extremely useful for debugging / understanding what's
going on with SCIP indices.  Note that it can take some time to run on
large indices and the output will be voluminous, so you'll always want
to pipe to a pager.

Originally we were updating to 0.2.3 but 0.3.0 makes the print command
also support JSON output via `--json` which seems potentially useful.