Remove generated rust checks; use SCIP analysis instead of rust.
It seems that the rust analyzer output does not include the generated
byproducts of the build script, although I think from the docs we
probably are letting rust-analyzer run it.  I am disabling the checks
by moving them to the side like we did for mozsearch; these can be
re-enabled if fixed in upstream.

Rebase note: Emilio removed these on trunk, so we remove them here
too.