Add scip_subtrees to config file and use it for test repo.
Rather than trying to do hacky things with finding files that end in
".scip", it's most straightforward to have the configuration file
explicitly name the SCIP files that the indexing process should
process as part of `scip-analyze.sh`.  For trees like m-c that want
to drive the ingestion themselves, they can do so and not list any
trees here.

This could be expanded in the future to help drive the build process,
but arguably the commands to generate the SCIP analysis should be
quite straightforward, so it's probably a reasonable convention for
the build files to do that themselves.