Move scip-indexer to use Config module for path info.
Although there is a use-case for the SCIP indexer to able to be told
paths independent of the canonical indexer paths for the purposes of
the m-c merge-analyses logic, it's arguably better to centralize any
logic related to that in `config.rs`.

To this end, this patch changes scip-indexer to use `config.rs` and
drop its own ad-hoc list of paths, most of which had become unused
code and were already removed.

This does mean if the code wants to support merge-analyses, it will
either be necessary to add some new logic.  I would suggest:

- Create a common clap struct that can be used with `#[clap(flatten)]`
  that individual commands can cite.  This allows centralizing the
  logic for any such diversions.
- Considering overhauling the merge-analyses shell script logic used
  for m-c to instead exist more inside rust logic in the first place.
  Ideally the shell scripts can be simpler now that we don't need to
  perform a bunch of hacky save-analysis fix-up passes for rust.